# docs/handoff/ups-expresosapp-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/routes": {
      "post": {
        "operationId": "RoutesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRouteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Route already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new route",
        "tags": [
          "Admin Routes"
        ]
      },
      "get": {
        "operationId": "RoutesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "description": "Page number (1-based)",
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "description": "Items per page",
            "required": false,
            "name": "limit",
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of routes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutePaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all routes with pagination",
        "tags": [
          "Admin Routes"
        ]
      }
    },
    "/admin/routes/{id}": {
      "get": {
        "operationId": "RoutesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Route ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Route not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get route details",
        "tags": [
          "Admin Routes"
        ]
      },
      "patch": {
        "operationId": "RoutesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Route ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRouteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Route not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a route",
        "tags": [
          "Admin Routes"
        ]
      }
    },
    "/admin/routes/{id}/stops/order": {
      "patch": {
        "operationId": "RouteStopsController_orderStops",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Route ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRouteStopsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stops ordered successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Route not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Order stops for a route",
        "tags": [
          "Admin Routes"
        ]
      }
    },
...
    "/admin/schedules": {
      "post": {
        "operationId": "SchedulesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schedule created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new schedule",
        "tags": [
          "Admin Schedules"
        ]
      },
      "get": {
        "operationId": "SchedulesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "routeId",
            "required": false,
            "in": "query",
            "description": "Filter by route ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "dayOfWeek",
            "required": false,
            "in": "query",
            "description": "Filter by day of week",
            "schema": {
              "example": "MONDAY",
              "type": "string",
              "enum": [
                "MONDAY",
                "TUESDAY",
                "WEDNESDAY",
                "THURSDAY",
                "FRIDAY",
                "SATURDAY",
                "SUNDAY"
              ]
            }
          },
          {
            "description": "Filter by route ID",
            "required": false,
            "name": "routeId",
            "in": "query",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "description": "Filter by day of week",
            "required": false,
            "name": "dayOfWeek",
            "in": "query",
            "schema": {
              "example": "MONDAY",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of schedules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchedulePaginatedResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List schedules with pagination and filters",
        "tags": [
          "Admin Schedules"
        ]
      }
    },
    "/admin/schedules/{id}": {
      "get": {
        "operationId": "SchedulesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Schedule ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get schedule details",
        "tags": [
          "Admin Schedules"
        ]
      },
      "patch": {
        "operationId": "SchedulesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Schedule ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedule updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Schedule not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a schedule",
        "tags": [
          "Admin Schedules"
        ]
      }
    },
...
    "/mobile/routes/{id}": {
      "get": {
        "operationId": "MobileController_findRouteDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Route ID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Route detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MobileRouteDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Route not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get route detail with ordered stops and active schedules",
        "tags": [
          "Mobile"
        ]
      }
    },
...
  },
  "info": {
    "title": "UPS ExpresosApp API",
    "description": "Backend API for UPS ExpresosApp MVP - Institutional transport management system for Universidad Politecnica Salesiana.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRouteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Route name",
            "example": "Norte - Salesiana"
          },
          "description": {
            "type": "string",
            "description": "Route description",
            "example": "Ruta que cubre el norte de la ciudad"
          },
          "direction": {
            "type": "string",
            "description": "Route direction",
            "example": "Norte"
          },
          "status": {
            "type": "string",
            "description": "Route status",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the route is active",
            "default": true
          }
        },
        "required": [
          "name",
          "direction"
        ]
      },
      "RouteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Norte - Salesiana"
          },
          "description": {
            "type": "object",
            "example": "Ruta que cubre el norte de la ciudad",
            "nullable": true
          },
          "direction": {
            "type": "string",
            "example": "Norte"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "direction",
          "status",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "RoutePaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateRouteDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Route name",
            "example": "Norte - Salesiana"
          },
          "description": {
            "type": "string",
            "description": "Route description",
            "example": "Ruta que cubre el norte de la ciudad"
          },
          "direction": {
            "type": "string",
            "description": "Route direction",
            "example": "Norte"
          },
          "status": {
            "type": "string",
            "description": "Route status",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the route is active",
            "default": true
          }
        }
      },
      "RouteStopOrderItemDto": {
        "type": "object",
        "properties": {
          "stopId": {
            "type": "string",
            "description": "Stop ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "stopOrder": {
            "type": "number",
            "description": "Order of the stop in the route",
            "example": 1
          },
          "estimatedArrivalMinutes": {
            "type": "number",
            "description": "Estimated arrival minutes from route start",
            "example": 15
          },
          "notes": {
            "type": "string",
            "description": "Notes for this stop in the route",
            "example": "Parada principal"
          }
        },
        "required": [
          "stopId",
          "stopOrder"
        ]
      },
      "OrderRouteStopsDto": {
        "type": "object",
        "properties": {
          "stops": {
            "description": "Ordered stops",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RouteStopOrderItemDto"
            }
          }
        },
        "required": [
          "stops"
        ]
      },
...
      "CreateScheduleDto": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "description": "Route ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dayOfWeek": {
            "type": "string",
            "description": "Day of week",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "direction": {
            "type": "string",
            "description": "Direction",
            "example": "Norte"
          },
          "departureTime": {
            "type": "string",
            "description": "Departure time in HH:mm format",
            "example": "07:30"
          },
          "approximateArrivalTime": {
            "type": "string",
            "description": "Approximate arrival time in HH:mm format",
            "example": "08:15"
          },
          "status": {
            "type": "string",
            "description": "Schedule status",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          }
        },
        "required": [
          "routeId",
          "dayOfWeek",
          "direction",
          "departureTime"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "routeId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "direction": {
            "type": "string",
            "example": "Norte"
          },
          "departureTime": {
            "type": "string",
            "example": "07:30"
          },
          "approximateArrivalTime": {
            "type": "object",
            "example": "08:15",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "routeId",
          "dayOfWeek",
          "direction",
          "departureTime",
          "approximateArrivalTime",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "SchedulePaginatedResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleResponseDto"
            }
          },
          "meta": {
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateScheduleDto": {
        "type": "object",
        "properties": {
          "routeId": {
            "type": "string",
            "description": "Route ID",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "dayOfWeek": {
            "type": "string",
            "description": "Day of week",
            "enum": [
              "MONDAY",
              "TUESDAY",
              "WEDNESDAY",
              "THURSDAY",
              "FRIDAY",
              "SATURDAY",
              "SUNDAY"
            ],
            "example": "MONDAY"
          },
          "direction": {
            "type": "string",
            "description": "Direction",
            "example": "Norte"
          },
          "departureTime": {
            "type": "string",
            "description": "Departure time in HH:mm format",
            "example": "07:30"
          },
          "approximateArrivalTime": {
            "type": "string",
            "description": "Approximate arrival time in HH:mm format",
            "example": "08:15"
          },
          "status": {
            "type": "string",
            "description": "Schedule status",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "ACTIVE"
          }
        }
      },
...
      "MobileRouteStopResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "stopOrder": {
            "type": "number",
            "example": 1
          },
          "estimatedArrivalMinutes": {
            "type": "object",
            "example": 15,
            "nullable": true
          },
          "notes": {
            "type": "object",
            "example": "Parada principal",
            "nullable": true
          },
          "stop": {
            "$ref": "#/components/schemas/StopResponseDto"
          }
        },
        "required": [
          "id",
          "stopOrder",
          "estimatedArrivalMinutes",
          "notes",
          "stop"
        ]
      },
      "MobileRouteDetailResponseDto": {
        "type": "object",
        "properties": {
          "route": {
            "$ref": "#/components/schemas/RouteResponseDto"
          },
          "stops": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MobileRouteStopResponseDto"
            }
          },
          "schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduleResponseDto"
            }
          }
        },
        "required": [
          "route",
          "stops",
          "schedules"
        ]
      },
...
    }
  }
}